requirements.txt
pymongo
pillow
numpy
imageio
